:sparkles: STring Enum값 잘못 기입할 시 에외
