+++ src/main/java/animal/meeting/global/error/handler/GlobalExceptionHandler.java
@@ -1,27 +1,21 @@
-package animal.meeting.global.error;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+package animal.meeting.global.error.handler;
 
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
 import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.NoHandlerFoundException;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
 
+import animal.meeting.global.error.ErrorCode;
 import animal.meeting.global.error.constants.CommonErrorCode;
 import animal.meeting.global.response.ErrorResponse;
 import animal.meeting.global.response.GlobalResponse;
-import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
 
 @RestControllerAdvice
@@ -38,6 +32,23 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 
 		ErrorResponse errorResponse = ErrorResponse.of(ex.getClass().getSimpleName(), ex.getMessage());
 		return super.handleExceptionInternal(ex, errorResponse, headers, statusCode, request);
+	}
+
+	/**
+	 * enum 타입 String의 값을 잘못입력 햇을 때
+	 */
+	@Override
+	protected ResponseEntity<Object> handleHttpMessageNotReadable(
+		HttpMessageNotReadableException ex,
+		HttpHeaders headers,
+		HttpStatusCode status,
+		WebRequest request) {
+
+		log.error("HttpMessageNotReadableException : {}", ex.getMessage(), ex);
+		final ErrorCode errorCode = CommonErrorCode.METHOD_ARGUMENT_TYPE_MISMATCH;
+		final ErrorResponse errorResponse = ErrorResponse.of(ex.getClass().getSimpleName(), errorCode.getMessage());
+		final GlobalResponse response = GlobalResponse.fail(errorCode.getHttpStatus().value(), errorResponse);
+		return ResponseEntity.status(errorCode.getHttpStatus()).body(response);
 	}
 
 	/**
@@ -109,68 +120,4 @@ public class GlobalExceptionHandler extends ResponseEntityExceptionHandler {
 		return ResponseEntity.status(errorCode.getHttpStatus()).body(response);
 	}
 
-	/** CustomException 예외 처리 */
-	@ExceptionHandler(CustomException.class)
-	public ResponseEntity<GlobalResponse> handleCustomException(CustomException ex) {
-
-		log.error("CustomException : {}", ex.getMessage(), ex);
-		final ErrorCode errorCode = ex.getErrorCode();
-		final ErrorResponse errorResponse = ErrorResponse.of(errorCode.name(), errorCode.getMessage());
-		final GlobalResponse response = GlobalResponse.fail(errorCode.getHttpStatus().value(), errorResponse);
-		return ResponseEntity.status(errorCode.getHttpStatus()).body(response);
-	}
-
-	/** 500번대 에러 처리 */
-	@ExceptionHandler(Exception.class)
-	protected ResponseEntity<GlobalResponse> handleAllException(Exception ex) {
-
-		log.error("Internal Server Error : {}", ex.getMessage(), ex);
-		final ErrorCode internalServerError = CommonErrorCode.INTERNAL_SERVER_ERROR;
-		final ErrorResponse errorResponse = ErrorResponse.of(ex.getClass().getSimpleName(), internalServerError.getMessage());
-		final GlobalResponse response = GlobalResponse.fail(internalServerError.getHttpStatus().value(), errorResponse);
-		return ResponseEntity.status(internalServerError.getHttpStatus()).body(response);
-	}
-
-
-
-	/** enum type 일치하지 않아 binding 못할 경우 발생 주로 @RequestParam enum으로 binding 못했을 경우 발생 */
-	@ExceptionHandler(MethodArgumentTypeMismatchException.class)
-	protected ResponseEntity<GlobalResponse> handleMethodArgumentTypeMismatchException(
-		MethodArgumentTypeMismatchException ex) {
-
-		log.error("MethodArgumentTypeMismatchException : {}", ex.getMessage(), ex);
-		final ErrorCode errorCode = CommonErrorCode.METHOD_ARGUMENT_TYPE_MISMATCH;
-		final ErrorResponse errorResponse = ErrorResponse.of(ex.getClass().getSimpleName(), errorCode.getMessage());
-		final GlobalResponse response = GlobalResponse.fail(errorCode.getHttpStatus().value(), errorResponse);
-		return ResponseEntity.status(errorCode.getHttpStatus()).body(response);
-	}
-
-	/** Request Param Validation 예외 처리 */
-	@ExceptionHandler(ConstraintViolationException.class)
-	public ResponseEntity<GlobalResponse> handleConstraintViolationException(
-		ConstraintViolationException e) {
-		log.error("ConstraintViolationException : {}", e.getMessage(), e);
-
-		Map<String, Object> bindingErrors = new HashMap<>();
-		e.getConstraintViolations()
-			.forEach(
-				constraintViolation -> {
-					List<String> propertyPath =
-						List.of(
-							constraintViolation
-								.getPropertyPath()
-								.toString()
-								.split("\\."));
-					String path =
-						propertyPath.stream()
-							.skip(propertyPath.size() - 1L)
-							.findFirst()
-							.orElse(null);
-					bindingErrors.put(path, constraintViolation.getMessage());
-				});
-
-		final ErrorResponse errorResponse = ErrorResponse.of(e.getClass().getSimpleName(), bindingErrors.toString());
-		final GlobalResponse response = GlobalResponse.fail(HttpStatus.BAD_REQUEST.value(), errorResponse);
-		return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(response);
-	}
 }
